# Increase test coverage