--- a/doc.go
+++ b/doc.go
@@ -1,4 +1,2 @@
-/*
-   Package toml is a library to read and write TOML documents.
-*/
+// Package toml is a library to read and write TOML documents.
 package toml