~Goal is 95%~ ✅ 97.6%

--- 

Some more allocations. Guessing this comes from the new array handling. Will get rid of them during the performance round.

```
name                              old time/op    new time/op    delta
UnmarshalDataset/config-32          78.6ms ± 4%    79.7ms ± 2%    ~     (p=0.165 n=10+10)
UnmarshalDataset/canada-32           105ms ± 2%     106ms ± 6%    ~     (p=0.356 n=9+10)
UnmarshalDataset/citm_catalog-32    55.6ms ± 6%    56.7ms ± 4%    ~     (p=0.143 n=10+10)
UnmarshalDataset/twitter-32         24.4ms ± 7%    24.9ms ± 5%    ~     (p=0.218 n=10+10)
UnmarshalDataset/code-32             294ms ± 4%     297ms ± 4%    ~     (p=0.113 n=10+9)
UnmarshalDataset/example-32          481µs ± 3%     475µs ± 4%    ~     (p=0.278 n=9+10)
UnmarshalSimple-32                  1.29µs ± 3%    1.27µs ± 4%    ~     (p=0.171 n=10+10)
ReferenceFile-32                    42.1µs ± 3%    40.4µs ± 3%  -3.99%  (p=0.000 n=9+9)

name                              old speed      new speed      delta
UnmarshalDataset/config-32        13.3MB/s ± 3%  13.2MB/s ± 2%    ~     (p=0.251 n=9+10)
UnmarshalDataset/canada-32        21.0MB/s ± 2%  20.8MB/s ± 6%    ~     (p=0.345 n=9+10)
UnmarshalDataset/citm_catalog-32  10.0MB/s ± 6%   9.8MB/s ± 4%    ~     (p=0.143 n=10+10)
UnmarshalDataset/twitter-32       18.1MB/s ± 7%  17.8MB/s ± 5%    ~     (p=0.197 n=10+10)
UnmarshalDataset/code-32          9.15MB/s ± 4%  9.03MB/s ± 4%    ~     (p=0.099 n=10+9)
UnmarshalDataset/example-32       16.9MB/s ± 3%  17.1MB/s ± 4%    ~     (p=0.278 n=9+10)
ReferenceFile-32                   125MB/s ± 3%   130MB/s ± 3%  +4.17%  (p=0.000 n=9+9)

name                              old alloc/op   new alloc/op   delta
UnmarshalDataset/config-32          16.9MB ± 0%    16.9MB ± 0%  +0.00%  (p=0.000 n=9+10)
UnmarshalDataset/canada-32          74.3MB ± 0%    74.3MB ± 0%  +0.00%  (p=0.000 n=10+10)
UnmarshalDataset/citm_catalog-32    37.1MB ± 0%    37.3MB ± 0%  +0.45%  (p=0.000 n=9+10)
UnmarshalDataset/twitter-32         15.6MB ± 0%    15.6MB ± 0%  +0.11%  (p=0.000 n=10+10)
UnmarshalDataset/code-32            59.3MB ± 0%    59.5MB ± 0%  +0.33%  (p=0.000 n=10+8)
UnmarshalDataset/example-32          238kB ± 0%     238kB ± 0%  +0.19%  (p=0.000 n=10+10)
ReferenceFile-32                    11.8kB ± 0%    11.8kB ± 0%    ~     (all equal)

name                              old allocs/op  new allocs/op  delta
UnmarshalDataset/config-32            645k ± 0%      645k ± 0%  +0.00%  (p=0.000 n=10+7)
UnmarshalDataset/canada-32            896k ± 0%      896k ± 0%  +0.00%  (p=0.000 n=10+10)
UnmarshalDataset/citm_catalog-32      370k ± 0%      380k ± 0%  +2.83%  (p=0.000 n=9+10)
UnmarshalDataset/twitter-32           157k ± 0%      158k ± 0%  +0.67%  (p=0.000 n=10+10)
UnmarshalDataset/code-32             2.91M ± 0%     2.92M ± 0%  +0.42%  (p=0.000 n=10+7)
UnmarshalDataset/example-32          3.63k ± 0%     3.66k ± 0%  +0.80%  (p=0.000 n=10+10)
ReferenceFile-32                       253 ± 0%       253 ± 0%    ~     (all equal)

```